+++ lambda/brotli-request.py
@@ -1,0 +1,13 @@
+import json
+
+def brotliRequest(event, context):
+  request = event['Records'][0]['cf']['request']
+  headers = request['headers']
+
+  if (headers['accept-encoding'] and 'br' in headers['accept-encoding'][0].value):
+    headers['x-compression'] = [{
+      'key': 'X-Compression',
+      'value': 'br'
+    }]
+
+  return request